requirements.dev.txt
blinker==1.8.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dash==2.17.0
dash-bootstrap-components==1.6.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
Flask==2.0.3
idna==3.7
importlib_metadata==7.1.0
itsdangerous==2.2.0
Jinja2==3.0.3
jinjasql==0.1.8
MarkupSafe==2.1.5
nest-asyncio==1.6.0
packaging==24.0
plotly==5.22.0
psycopg2==2.9.9
python-dotenv==1.0.1
requests==2.32.3
retrying==1.3.4
setuptools==70.0.0
six==1.16.0
tenacity==8.3.0
typing_extensions==4.12.0
urllib3==2.2.1
waitress==3.0.0
Werkzeug==2.2.2
zipp==3.19.0
